Fix warnings from ST compiler

+++ modules/libraries/standard/common/test/iot_test_platform_threads.c
@@ -145,7 +145,7 @@ TEST( UTIL_Platform_Threads, IotThreads_ThreadPriority )
     }
 
     TEST_ASSERT_EQUAL( 0, attrData );
-    printf( "Expected Pri = 0, actual = %d\r\n", attrData );
+    printf( "Expected Pri = 0, actual = %d\r\n", ( int ) attrData );
     attrData = -1;
 
     /* Create thread with priority 5 */
@@ -157,7 +157,7 @@ TEST( UTIL_Platform_Threads, IotThreads_ThreadPriority )
         vTaskDelay( 1 );
     }
 
-    printf( "Expected Pri = 5, actual = %d\r\n", attrData );
+    printf( "Expected Pri = 5, actual = %d\r\n", ( int ) attrData );
     TEST_ASSERT_EQUAL( 5, attrData );
     attrData = -1;
 
@@ -170,7 +170,7 @@ TEST( UTIL_Platform_Threads, IotThreads_ThreadPriority )
         vTaskDelay( 1 );
     }
 
-    printf( "Expected Pri = 7, actual = %d\r\n", attrData );
+    printf( "Expected Pri = 7, actual = %d\r\n", ( int ) attrData );
 }
 /*-----------------------------------------------------------*/
